requirements.txt
Flask
langchain
langchain-google-genai
google-generativeai
pypdf
faiss-cpu
python-dotenv
tiktoken
langchain-community
werkzeug
